More removal of warnings.

+++ src/SAX2srcMLHandler.cpp
@@ -179,7 +179,7 @@ void startRoot(void * ctx, const xmlChar * localname, const xmlChar * prefix, co
     state->root.attributes[index + 2] = attributes[index + 2] ? (xmlChar*) strdup((const char*) attributes[index + 2]) : 0;
     CHECK_COPY(attributes[index + 2], state->root.attributes[index + 2]);
 
-    int vallength = attributes[index + 4] - attributes[index + 3];
+    int vallength = (int)(attributes[index + 4] - attributes[index + 3]);
     state->root.attributes[index + 3] = (const xmlChar*) malloc(vallength);
     CHECK_COPY(attributes[index + 3], state->root.attributes[index + 3]);
 
@@ -241,13 +241,13 @@ void startElementNsFirst(void * ctx, const xmlChar * localname, const xmlChar * 
     state->process->startUnit(state->root.localname, state->root.prefix, state->root.URI,
                                state->root.nb_namespaces, state->root.namespaces, state->root.nb_attributes,
                                state->root.nb_defaulted, state->root.attributes);
-    state->process->charactersUnit((const xmlChar *)state->root.characters.c_str(), state->root.characters.size());
+    state->process->charactersUnit((const xmlChar *)state->root.characters.c_str(), (int)state->root.characters.size());
     state->process->startElementNs(localname, prefix, URI,
                                nb_namespaces, namespaces, nb_attributes,
                                nb_defaulted, attributes);
   } else {
 
-    state->process->charactersRoot((const xmlChar *)state->root.characters.c_str(), state->root.characters.size());
+    state->process->charactersRoot((const xmlChar *)state->root.characters.c_str(), (int)state->root.characters.size());
     state->process->startUnit(localname, prefix, URI,
                                nb_namespaces, namespaces, nb_attributes,
                                nb_defaulted, attributes);
@@ -387,7 +387,7 @@ void endElementNs(void * ctx, const xmlChar * localname, const xmlChar * prefix,
                                 state->root.nb_defaulted, state->root.attributes);
 
       if(state->root.characters.size() != 0)
-	state->process->charactersUnit((const xmlChar *)state->root.characters.c_str(), state->root.characters.size());
+	state->process->charactersUnit((const xmlChar *)state->root.characters.c_str(), (int)state->root.characters.size());
 
     }
